chore(experiments): extracting the breakdown code into a new class for readability.
